change cleanDatabase function

+++ src/utils/cleanDatabase.ts
@@ -1,8 +1,12 @@
 import Score from "../models/Score";
 
 export default async function cleanDatabase() {
-    const scores = await Score.find({ beatLeaderData: null });
+    const scores = await Score.find();
+    let i = 0;
     for (const score of scores) {
-        Score.deleteOne({ _id: score._id });
+        if (!score.beatLeaderData || !score.beatLeaderData.timeset) {
+            Score.deleteOne({ _id: score._id });
+        }
     }
+    console.log("Removed " + i + "invalid scores!");
 }
